DisasterSAS/disaster_backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
flask-jwt-extended
flask-socketio
psycopg2-binary
requests
werkzeug
twilio
apscheduler
